Upgraded to Storybook v6.1
