+++ .storybook/main.js
@@ -1,3 +1,10 @@
 module.exports = {
-    addons: ['@storybook/addon-viewport/register'],
-};
+  "stories": [
+    "../src/**/*.stories.mdx",
+    "../src/**/*.stories.@(js|jsx|ts|tsx)"
+  ],
+  "addons": [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials"
+  ]
+}
